Lamentations: start use of git.
The project began in March 2020.  Over time it became clearer that I should have used git from the outset.  But I hadn't done so.

This, two years later, is the beginnings of git to track it.

It isn't the earliest version, of course.  But at least I had kept hold of the versions prior to Adam Carlill's excellent suggestions for chapters one and two, so this first commit is a rough approximation to the state prior to that for those chapters.