chap2:4,6: adjust the two "tent" nouns and environs
The nouns need to be different, with that in v6 being more flimsy than that in v4.

While the previous (4:campsite; 6:tent) had this, nevertheless most translations use "tent" for v4.

So for v6 adopt Goldingay's "bivouac", allowing v4 to revert to "tent".

Also change:
* Felling his... like a garden
to:
* Flattening his... like a garden

which makes more obvious senes, while maintaining the acrostic "f".

Add footnote to v6.