2:15: complete rewrite
YLT (Young's Literal Translation), reflecting the Hebrew, indicates desirablity for a far stronger opening verb.
And chapter 2 is characterised by verb-fronted verses, with strong verbs.

This requires a re-ordering of the verse.  This was already out of order; this change improves that somewhat.

It also allows other improvements in its angst.