various: change "holocaust" to "ravaging"
Discover that 2:13 uses the same Hebrew word "seber" that I had previously represented as "holocaust",
so we should try to include that commonality.

But it doesn't really work there.

Further, the previous appropriation of the word "holocaust" had always worried me a little,
as imposing perhaps too much of a modern slant.