5:1: better punctuation
The two halves need better separation than a comma. But a fullstop is too much. And the semicolon in the second half should probably remain.

So use a colon. This also is subliminally more psalm-like.

I realise that the second half also ends with a colon.  But that is serving the different function of introducing the list that follows it.