notes: remove statements that are no longer significant
In the early days, being largely focussed on acrostics+qinah,
I had underplayed attention to details of words, and had given
more elbow-room to translations of words.

But the project evolved and began also to recognise some of the
important detail that had been getting unintentionally lost.

This part of the "notes" had reflected that earlier approach;
it is no longer relevant.